refactor: move multistore to separate folder
