+++ core/lib/multistore/index.ts
@@ -1,15 +1,14 @@
-import rootStore from '../store'
+import rootStore from './../../store'
 import { loadLanguageAsync } from '@vue-storefront/i18n'
-import { initializeSyncTaskStorage } from './sync/task'
+import { initializeSyncTaskStorage } from './../sync/task'
 import { Logger } from '@vue-storefront/core/lib/logger'
-import Vue from 'vue'
 import queryString from 'query-string'
 import merge from 'lodash-es/merge'
 import VueRouter, { RouteConfig, RawLocation } from 'vue-router'
 import config from 'config'
 import { coreHooksExecutors } from '@vue-storefront/core/hooks'
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
-import { LocalizedRoute, StoreView } from './types'
+import { LocalizedRoute, StoreView } from './../types'
 import storeCodeFromRoute from './storeCodeFromRoute'
 import cloneDeep from 'lodash-es/cloneDeep'
 import get from 'lodash-es/get'
